+++ lib/db_mod/statements/configuration/as/csv.rb
@@ -7,17 +7,14 @@ module DbMod
         # May be enabled for a prepared method or
         # statement method using +.as(:csv)+:
         #
-        #  def_statement(:a, 'SELECT a, b FROM foo').as(:csv)
-        #  def_prepared(:b, 'SELECT b, c FROM bar').as(:csv)
+        #  def_statement(:a, 'SELECT a, b FROM foo') { as(:csv) }
+        #  def_prepared(:b, 'SELECT b, c FROM bar') { as(:csv) }
         #
         #  def do_stuff
         #    a # => "a,b\r\n1,2\r\n3,4\r\n..."
         #  end
-        module Csv
-          # Enables this module to be passed to
-          # {DbMod::Statements::Configuration.process_method_results} as the
-          # +wrapper+ function, in which case it will retrieve the results
-          # and format them as a CSV document using the column names
+        class Csv
+          # Formats the results as a CSV document using the column names
           # from the result set.
           #
           # @param results [Object] SQL result set