+++ lib/db_mod/statements/configuration/as/json.rb
@@ -7,24 +7,22 @@ module DbMod
         # May be enabled for a prepared method or
         # statement method using +.as(:json)+:
         #
-        #  def_statement(:a, 'SELECT a, b FROM foo').as(:json)
-        #  def_prepared(:b, 'SELECT b, c FROM bar').as(:json)
+        #  def_statement(:a, 'SELECT a, b FROM foo') { as(:json) }
+        #  def_prepared(:b, 'SELECT b, c FROM bar') { as(:json) }
         #
         #  def do_stuff
         #   a # => '[{"a":"x","b":"y"},...]'
         #  end
-        module Json
-          # Enables this module to be passed to
-          # {DbMod::Statements::Configuration.process_method_results} as the
-          # +wrapper+ function, in which case it will retrieve the results
-          # and format them as a JSON string using the column names
-          # from the result set for the keys of each object.
+        class Json
+          # Formats the SQL results as a JSON object.
           #
           # @param results [Object] SQL result set
           # @return [String] a JSON formatted string
           def self.call(results)
-            # .map turns the result object into an array
-            results.map { |x| x }.to_json
+            # For compatibility with single(:row)
+            return results.to_json if results.is_a? Hash
+
+            results.to_a.to_json
           end
         end
       end