+++ lib/db_mod/statements/configuration/single.rb
@@ -13,14 +13,14 @@ module DbMod
       # is via {ConfigurableMethod#single}, which is available
       # when defining a statement method or prepared method:
       #
-      #  def_statement(:a, 'SELECT name FROM a WHERE id=$1').single(:value)
-      #  def_prepared(:b, 'SELECT id FROM b WHERE value > $min').single(:column)
-      #  def_prepared(:c, 'SELECT * FROM c WHERE id = $id').single(:row)
+      #  def_statement(:a, 'SELECT name FROM a WHERE id=$1') { single(:value) }
+      #  def_prepared(:b, 'SELECT id FROM b WHERE x > $y') { single(:column) }
+      #  def_prepared(:c, 'SELECT * FROM c WHERE id = $id') { single(:row) }
       #
       #  def do_stuff
-      #    a # => "foo"
-      #    b # => ['1','2','3',...]
-      #    c # => Hash
+      #    a(1)    # => "foo"
+      #    b(y: 2) # => ['1','2','3',...]
+      #    c id: 3 # => Hash
       #  end
       #
       # +.single(:row)+ and +.single(:value)+ will return the first
@@ -30,9 +30,6 @@ module DbMod
       # instead of returning +nil+, use +.single(:row!)+ or
       # +.single(:value!)+.
       module Single
-        # For process_method_results
-        Configuration = DbMod::Statements::Configuration
-
         # List of allowed parameters for {#single},
         # and the methods used to process them.
         COERCERS = {
@@ -42,21 +39,6 @@ module DbMod
           row!: Single::RequiredRow,
           column: Single::Column
         }
-
-        # Extend a method so that only some singular part of
-        # the SQL result set is returned.
-        # See above for more details.
-        #
-        # @param mod [Module] module where the method has been defined
-        # @param name [Symbol] method name
-        # @param type [Symbol] one of {SINGLE_TYPES}
-        def self.extend_method(mod, name, type)
-          unless COERCERS.key? type
-            fail ArgumentError, "#{type} not in #{COERCERS.keys.join ', '}"
-          end
-
-          Configuration.process_method_results(mod, name, COERCERS[type])
-        end
       end
     end
   end