+++ lib/db_mod/statements/statement.rb
@@ -61,12 +61,12 @@ module DbMod
       # @param mod [Module] a module with {DbMod} included
       def self.define_def_statement(mod)
         mod.class.instance_eval do
-          define_method(:def_statement) do |name, sql|
+          define_method(:def_statement) do |name, sql, &block|
             sql = sql.dup
             name = name.to_sym
 
             params = Parameters.parse_params! sql
-            Statement.define_statement_method(mod, name, params, sql)
+            Statement.define_statement_method(mod, name, params, sql, &block)
           end
         end
       end
@@ -80,62 +80,15 @@ module DbMod
       # @param params [Fixnum,Array<Symbol>]
       #   expected parameter count, or a list of argument names.
       #   An empty array produces a no-argument method.
-      def self.define_statement_method(mod, name, params, sql)
-        if params.is_a?(Array)
-          if params.empty?
-            define_no_args_statement_method(mod, name, sql)
-          else
-            define_named_args_statement_method(mod, name, params, sql)
-          end
+      # @yield dsl block may be passed, which will be evaluated using a
+      #   {MethodConfiguration} object as scope
+      def self.define_statement_method(mod, name, params, sql, &block)
+        if params == []
+          Configuration.def_configurable(mod, name, ->(*) { query sql }, &block)
         else
-          define_fixed_args_statement_method(mod, name, params, sql)
+          method = ->(*args) { conn.exec_params(sql, args) }
+          Configuration.def_configurable(mod, name, method, params, &block)
         end
-      end
-
-      # Define a no-argument method with the given name
-      # that will execute the given sql statement and return
-      # the result.
-      #
-      # @param mod [Module] {DbMod} enabled module
-      #   where the method will be defined
-      # @param name [Symbol] name of the method to be defined
-      # @param sql [String] parameterless SQL statement to execute
-      def self.define_no_args_statement_method(mod, name, sql)
-        Configuration.def_configurable mod, name, ->() { query(sql) }
-      end
-
-      # Define a method with the given name, that accepts the
-      # given set of named parameters that will be used to execute
-      # the given SQL query.
-      #
-      # @param mod [Module] {DbMod} enabled module
-      # @param name [Symbol] name of the method to be defined
-      # @param params [Array<Symbol>] parameter names and order
-      def self.define_named_args_statement_method(mod, name, params, sql)
-        method = lambda do |*args|
-          args = Parameters.valid_named_args! params, args
-          conn.exec_params(sql, args)
-        end
-
-        Configuration.def_configurable mod, name, method
-      end
-
-      # Define a method with the given name that accepts a fixed number
-      # of arguments, that will be used to execute the given SQL query.
-      #
-      # @param mod [Module] {DbMod} enabled module where the method
-      #   will be defined
-      # @param name [Symbol] name of the method to be defined
-      # @param count [Fixnum] arity of the defined method,
-      #   the number of parameters that the SQL statement requires
-      def self.define_fixed_args_statement_method(mod, name, count, sql)
-        method = lambda do |*args|
-          Parameters.valid_fixed_args!(count, args)
-
-          conn.exec_params(sql, args)
-        end
-
-        Configuration.def_configurable mod, name, method
       end
     end
   end