+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+0.0.5
+=====
+Breaking changes. Any statement or prepared methods with additional configuration
+should have such configuration declared in a block, rather than a method chain
+(although methods can be chained inside the method block). E.g.
+
+```ruby
+def_statement(:a, 'SELECT * FROM foo WHERE id = $1').single(:row).as(:json)
+
+# ... becomes ...
+
+def_statement(:a, 'SELECT * FROM foo WHERE id = $1') { single(:row).as(:json) }
+```
+
+* +def_statement+ and +def_prepared+ method configuration must now be supplied
+  as a block. This allows method configuration to be collected ahead of
+  method definition, and hence a bit more smarts during the method declaration
+  process - [@dslh](https://github.com/dslh).
+* `single(:row/:column).as(:json)` now works - [@dslh](https://github.com/dslh).
+
 0.0.4 (2015-10-15)
 ==================